RFCT: move get image file response for easier maintain
